v0.2.1 - Add remainder of skill tables

+++ race/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Race(models.Model):
     name = models.CharField(unique=True, max_length=32)
-    attribute_points = models.PositiveIntegerField()
+    attribute_points = models.PositiveSmallIntegerField()
     description = models.TextField()
     costume_req = models.TextField(blank=True, null=True)
     costume_sugg = models.TextField(blank=True, null=True)
@@ -18,7 +18,7 @@ class Race(models.Model):
 class SubRace(models.Model):
     name = models.CharField(unique=True, max_length=32)
     race = models.ForeignKey(Race, models.RESTRICT)
-    attribute_points = models.PositiveIntegerField()
+    attribute_points = models.PositiveSmallIntegerField()
     description = models.TextField()
     costume_req = models.TextField()
     costume_sugg = models.TextField()
